# Improve performance of `slice::sort` with ipn_stable