This is achieved by writing code that exploits instruction-parallel hardware:

- New small sort, that handles anything up to size 32
- Faster bi-directional merge function
- Faster run creation with the help of a sorting-netwok

Probing for a common value as done in ipn_stable is omitted and can be added incrementally in the future. A detailed writeup can be found here https://github.com/Voultapher/sort-research-rs/blob/main/writeup/glidesort_perf_analysis/text.md. Only difference being low-cardinality inputs, which will perform similar to random inputs.