@Voultapher 2023-02-14
I developed a pretty thorough test suite https://github.com/Voultapher/sort-research-rs/blob/main/tests/main.rs, that checks things like, will the original elements still be in the input if any of the possible comparison calls panic, what if the type has self-modified on the stack via interior mutability and the comparison function panics, will all modifications be visible. Various ways in which Ord can be violated and more. All that with various input sizes and patterns. I run this test suite normally, with ASAN or with miri. I have a libfuzzer and a afl harnesses. This version of the code passes all those tests.