@thomcc 2023-02-19
This looks pretty good to me. It's certainly complex, but is very well commented.

@bors try @rust-timer queue