@Voultapher 2023-02-19
Interesting regression results. I could see it being a real slowdown for smallish inputs size < 15, and I have some ideas how to address that. At the same time I wonder how representative that metric is. This new implementation deliberately exploits instruction-level parallelism. So it certainly issues and retires more instructions, but is faster in terms of wall clock. Is there a way to have these measurements done with time instead of instruction count?

Tagging @rylev, you seem to know a lot about these performance regression benchmarks.