@Voultapher 2023-02-19
@thomcc I've pushed a new commit 'Re-tune ipn_stable small sort for cold code', that greatly improves performance for small inputs in cold code. Details found in the commit message and the code.

@bors try @rust-timer queue