@thomcc 2023-02-19
> This new implementation deliberately exploits instruction-level parallelism

Note that it's a regression in cycles too, which would usually be reflected. It's possible that the small size issues are the cause though.

@bors try @rust-timer queue