@Voultapher 2023-02-19
> Note that it's a regression in cycles too

Good point, I missed that. Yes cycles ought to go down.