@thomcc 2023-02-20
I'm a bit concerned. There's a chance the slowdown is coming from there being more code to monomorphize for each call to sort, but it's not clear that that's the case.