@Voultapher 2023-02-20
You might be onto something. The heuristic functions are for technical reasons non-const, so it has to generate all the code, eg. the stable sorting-network for all types regardless if they qualify. I have some ideas how to address that. Will update.