@Voultapher 2023-02-20
I've pruned the code in a way that greatly reduces produced LLVM-IR for debug and release builds. One sacrifice I made, was the small sort with analysis, as consequence the performance for slices in the size range 16-40 is worse than before. But I think that's a worthwhile sacrifice for smaller code-size and faster build times.