@Voultapher 2023-02-20
Not ideal yet, but progress. Let me drop the sorting-network from 8 down to 6, that should improve comparison efficiency and reduce code-size even more.