@Voultapher 2023-02-20
After some more pruning I'm now down to:

LLVM-IR generated for single instantiation of `slice::sort` for debug build:

- `i32` 703kB -> 1417kB (include bi-directional merge and sorting-network)
- `String` 1043kB -> 1180kB (includes bi-directional merge)

I tried going form a 8 element sorting-network down to a more comparison efficient 6 element one, but that yields a 5-10% performance degradation, even for expensive to compare types, `f128` in my benchmarks. So I'm not convinced that's worth it.

I've pushed 'Use correct has_direct_iterior_mutability implementation' which should yield speedups for types such as `String` for slices size > 20. Maybe that can counteract some of the regression caused by more code being generated.