@Voultapher 2023-02-20
Note, in my testing the `if const { const_fn_xx::<T>() } { ... }` did not reliably prune the code from being generated as LLVM-IR, to me this seems like a bug, but the effect is minor here as the big expensive part is safe-guarded by trait specialization. The code is written in a way that should enable even smaller code-footprints in the future.