@thomcc 2023-02-20
I'll need to re-review the specialization parts.