@thomcc 2023-02-21
To be clear, I'm not sure that the specialization is the explanation entirely. This has a lot of (fairly consistent) regressions, across both primary and secondary benchmarks, and I don't think all of them have sorts in them.