@Voultapher 2023-02-21
Looking at the cycle metric, there are significantly fewer regressions than in the instruction metric, which would match the higher instructions/cycle with ipn_stable. A mean compile time regression of 1.5% seems an ok price to pay for a ~2x speedup of the core algorithm for integer like types, and a smaller but noticeable ~1.35x speedup for types like `String`.