@cuviper 2023-02-21
It could be evidence that the compiler's own sorting performance is negatively impacted.