@Voultapher 2023-02-23
I ran some benchmarks in real world scenarios and got a little spooked by the stable sorting-network, it's great for integers but I suspect that limiting it to `Copy` is not enough and I saw some serious regressions. So to be on the safe side and to limit code-gen even further completely deleted it. While a lot less ambitious, the current version still yields:

```
hot-u64-random-10k

with network and bi-dir merge: 110us
with bi-dir merge: 165us (current PR state)
master: 260us

hot-string-random-10k

with bi-dir merge: 848us
master: 1214us
```

Would be good do another timing run.

Also arguably, for integers users should use unstable sort, which I plan on upstreaming changes to soon (tm).